Images load w/o pressing home bttn & some refactoring
